% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STV-package.R
\docType{data}
\name{ballots}
\alias{ballots}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
John W. Emerson \email{john.emerson@yale.edu}
}
\keyword{data}
