% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_pacf.R
\name{stv_pacf}
\alias{stv_pacf}
\title{Gráfico da Função de Autocorrelação Parcial de uma Série Temporal.}
\usage{
stv_pacf(
  vetor,
  cor = "purple",
  bolinha = FALSE,
  linha = 1,
  lagg = 20,
  tema = "claro"
)
}
\arguments{
\item{vetor}{Vetor contendo os valores da série. Precisa ser numérica.}

\item{cor}{String contendo a cor do gráfico, podendo ser no formato hexadecimal.}

\item{bolinha}{Argumento do tipo lógico. \code{bolinha = TRUE} retorna o PACF em um gráfico lollipop.}

\item{linha}{Valor numérico indicando a expessura da linha.}

\item{lagg}{Defasagem máxima para o cálculo do ACF.}

\item{tema}{String contendo o tema do gráfico. Precisa ser necessariamente 'claro' ou 'escuro'.}
}
\value{
Um gráfico da função de autocorrelação parcial de uma série temporal.
}
\description{
Gráfico da Função de Autocorrelação Parcial de uma Série Temporal.
}
\examples{
library(stv)
library(ggdark)
library(ggplot2)
teste <- data.frame(data = seq(1,100), valor = rnorm(100))
stv_pacf(teste$valor)
stv_pacf(teste$valor, "magenta", TRUE)
stv_pacf(teste$valor, linha = 2, lag = 10, tema = "escuro")

}
