% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_seq.R
\name{stv_seq}
\alias{stv_seq}
\title{Gráfico de Sequência de Séries Temporais}
\usage{
stv_seq(
  dados,
  x,
  y,
  eixo_x = x,
  eixo_y = y,
  titulo = "Gráfico de Sequência",
  bolinha = FALSE,
  cor = "purple",
  tema = "claro",
  anim = FALSE
)
}
\arguments{
\item{dados}{Objeto do tipo \code{Data.Frame} contendo os dados.}

\item{x}{String contendo o nome de uma coluna de \code{dados} indicando data ou índice. A coluna precisa ser numérica.}

\item{y}{String contendo o nome de uma coluna de \code{dados} indicando valores da série. A coluna precisa ser numérica.}

\item{eixo_x}{String contendo o nome do eixo x.}

\item{eixo_y}{String contendo o nome do eixo y.}

\item{titulo}{String contendo o título do gráfico.}

\item{bolinha}{Argumento do tipo lógico indicando a inclusão pontos.}

\item{cor}{String contendo a cor do gráfico, podendo ser no formato hexadecimal.}

\item{tema}{String contendo o tema do gráfico. Precisa ser necessariamente 'claro' ou 'escuro'.}

\item{anim}{Argumento do tipo lógico indicando a inclusão de animação.}
}
\value{
Um gráfico de sequência da série temporal.
}
\description{
Gráfico de Sequência de Séries Temporais
}
\examples{
library(stv)
library(ggdark)
library(ggplot2)
teste <- data.frame(data = seq(1,100), valor = rnorm(100))
stv_seq(teste, "data", "valor")
stv_seq(teste, "data", "valor", "Índice", "Valor da Série", bolinha = TRUE, tema = "escuro")
stv_seq(teste, "data", "valor", cor = "magenta", anim = TRUE)

}
