% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_tabela.R
\name{stv_tabela}
\alias{stv_tabela}
\title{Tabela das Principais Estatísticas}
\usage{
stv_tabela(dados)
}
\arguments{
\item{dados}{Vetor numérico d contendo os valores da série.}
}
\value{
Uma tabela das principais estatísticas da série temporal.
}
\description{
Função que retorna uma tabela com as principais estatísticas da série temporal
}
