% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STV-package.R
\docType{data}
\name{ballots}
\alias{ballots}
\title{ballots data contains rankings of 44 candidates by 489 people.
10 rows are black, which will prompt an error. This data follows the structure of a 2016 Yale faculty senate election, but vote choices have been permuted and candidate names have been removed.}
\description{
ballots data contains rankings of 44 candidates by 489 people.
10 rows are black, which will prompt an error. This data follows the structure of a 2016 Yale faculty senate election, but vote choices have been permuted and candidate names have been removed.
}
\author{
John W. Emerson \email{john.emerson@yale.edu}
}
\keyword{data}
