% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_clean_ballots.R
\name{cleanBallots}
\alias{cleanBallots}
\title{Tries to Address \code{validateBallots()} Warning(s).}
\usage{
cleanBallots(x, cand.names = NULL)
}
\arguments{
\item{x}{a \code{data.frame} with rows as ballots and columns as candidates.}

\item{cand.names}{character vector of length equal to number of candidates
(needed when column names of \code{x} are missing or not unique).}
}
\value{
a \code{data.frame} compatible for \code{stv()} function.
}
\description{
Tries to clean data for \code{stv()}. Some of the warnings from \code{validateBallots()}
have to addressed by the user (see Details).
}
\details{
Assumes \code{x} contains rows and columns corresponding to ballots and
candidates respectively. Tries to address issues raised by \code{validateBallots()}
in the following order:
\enumerate{
  \item If \code{x} is a \code{matrix} then converts to \code{data.frame}.
    Otherwise, user has to convert \code{x} into \code{data.frame}.
  \item Checks if \code{x} has numeric entries. If not, checks if numeric
    data was passed as character. If this also fails, then user has to
    convert data into numeric type.
  \item If column names of \code{x} are missing assigns \code{cand.names} as
    column names. If
  \code{x} already has valid column names, no need to specify \code{cand.names}.
  If column names of \code{x} missing and \code{cand.names} not specified, returns
  error message.
  \item Removes blank columns.
  \item Removes blank and/or non-sequentially ranked rows.
}
}
\examples{
data(ballots)
cballots <- cleanBallots(ballots)
validateBallots(cballots)
}
