% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STV-package.R
\docType{package}
\name{STV-package}
\alias{STV-package}
\title{STV Vote Counting}
\description{
Functions in this package are designed to clean, validate, and count single transferable votes.
}
\author{
Jay Emerson (expand this)
}
