% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{Implement STV counting system.}
\usage{
stv(x, seats = 1, file = "", surplusMethod = "Cambridge",
  quotaMethod = "Droop")
}
\arguments{
\item{x}{a data.frame with rows as ballots and columns as candidates. \code{x}
must pass all checks from \code{validateBallots()}.}

\item{seats}{a number indicating candidates to elect. (default = 1)}

\item{file}{a character string naming file. "" indicates output to the console only (default).
Saves a CSV file. So, name should contain ".csv" at the end.}

\item{surplusMethod}{a character string indicating which method to use for
surplus allocation. Currently supports only one option "Cambridge" (default)}

\item{quotaMethod}{a character string indicating which method to use for
calculation of quota. Currently support only one option "Droop" (default)}
}
\value{
a data.frame with rows containing detailed results from each round of STV counting.
    For any given round of counting, a row contains: number of active ballots, seats to fill,
    quota, maximum and minimum votes obtained by any candidate, who was eliminated (if any),
    if there was a tie for elimination (if yes, how many tied), who was elected (if any),
    surplus if elected and each candidate's votes tally for that round.
}
\description{
\code{stv} returns a data.frame with rows containing detailed results
from each round of STV counting.
}
\details{
\code{stv()} first validates \code{x} by running \code{validateBallots()} function.
Once validation is complete, it implements Single Transferable Vote
counting method. Each round of counting starts with idetification of active
ballots. Then quota is calculated (currently only supports Droop method \code{INCLUDE FORMULA?}).
Tally of candidates is obtained using top choices of active ballots. If a
candidate reaches quota, she/he is elected and associated surplus ballots
are reallocated (currently only supports Cambridge method \code{ONE LINE DETAIL?}). If multiple candidates
reach quota, all of them are elected and their surplus are allocated in that
round itself. If nobody reaches quota then candidate with minimum number of votes
is eliminated. If multiple candidates tie for minimum number of votes, one of
them is selected at random and eliminated. The process is repeated till all
of the seats are filled or number of candidates still in race equals number
of unfilled seats. In later case, all of the active candidates are elected.
Note: a ballot stays active till either it runs out of marked choices or gets
removed during surplus allocation.
}
\examples{
#stv(voting.data, seats = 4, file = "voting_results.csv")
}
