% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_seq.R
\name{stv_seq}
\alias{stv_seq}
\title{Gráfico Sequencial}
\usage{
stv_seq(
  dados,
  x,
  y,
  eixo_x = x,
  eixo_y = y,
  titulo = "Gráfico de Sequência",
  bolinha = FALSE,
  cor = "purple",
  tema = "claro",
  anim = F
)
}
\arguments{
\item{dados}{objeto do tipo data.frame;}

\item{x}{nome da coluna no data frame dos dados do eixo x}

\item{y}{nome da coluna no data frame dos dados do eixo y}

\item{eixo_x}{denominacao no eixo x, default é o nome da coluna informado no parametro x}

\item{eixo_y}{denominacao no eixo y, default é o nome da coluna informado no parametro y}

\item{bolinha}{gr?fico ter? as observa??es marcadas por bolinhas ou n?o, default ? sem bolinha (boleano);}

\item{cor}{cor da linha e das bolinhas (character), default ? preto;}

\item{tema}{tema do gráfico, default = claro}

\item{anim}{se o grqfico sera animado ou nao (boleano).}

\item{t?tulo}{objeto do tipo character, default = "Grafico de Sequencia";}
}
\value{
Gráfico Sequencial
}
\description{
Função que retorna um gráfico ACF de uma série temporal passada por parâmetro
}
