% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_tabela.R
\name{stv_tabela}
\alias{stv_tabela}
\title{Tabela das principais estatísticas}
\usage{
stv_tabela(dados)
}
\arguments{
\item{dados}{dados da série temporal}
}
\value{
Data Frame
}
\description{
Função que retorna uma tabela com as principais estatísticas da série temporal
}
