% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_clean_ballots.R
\name{validateBallots}
\alias{validateBallots}
\title{Validates Input Data Format for \code{stv()}.}
\usage{
validateBallots(x)
}
\arguments{
\item{x}{a \code{data.frame} with rows as ballots and columns as candidates.}
}
\value{
string: "All tests passed. Please feel free to run stv() function." if \code{x}
  passes all tests. Otherwise, corresponding warning message. NOTE: Some of the
  warnings can be fixed using \code{cleanBallots()} function. Other issues must be
  fixed by the user.
}
\description{
Validates data format for \code{stv()} input. Assumes each row and column
corresponds to a ballot and candidate respectively. See details for the
tests run by this function. If input data is in correct format, returns
string: "All tests passed. Please feel free to run \code{stv()} function.".
Otherwise, prints corresponding warning message.
}
\details{
Checks if input data is in acceptable format for \code{stv()}. User must run
this function before calling \code{stv()}. Before any calculation, \code{stv()}
calls this function and proceeds only if \code{x} passes all of the following tests
(run in given order).
\enumerate{
  \item If \code{x} is a \code{data.frame}.
  \item If \code{x} has unique column names.
  \item If \code{x} has numeric entries.
  \item If \code{x} doesn't have any blank column.
  \item If \code{x} doesn't have any blank row or a row with non-sequential ranks.
}
}
\examples{
data(ballots)
result <- try(validateBallots(ballots), silent=TRUE)
print(result)
}
